feat: separate gig buyer profile
