+++ actions/user/gig-buyer-profile.ts
@@ -1,169 +1,16 @@
 "use server";
 
 import { db } from "@/lib/drizzle/db";
-import {
-  BadgeDefinitionsTable,
-  BuyerProfilesTable,
-  GigsTable,
-  PaymentsTable,
-  ReviewsTable,
-  UserBadgesLinkTable,
-  UsersTable,
-} from "@/lib/drizzle/schema";
+import { BuyerProfilesTable, UsersTable } from "@/lib/drizzle/schema";
 import { ERROR_CODES } from "@/lib/responses/errors";
 import { isUserAuthenticated } from "@/lib/user.server";
-import { and, eq, gt } from "drizzle-orm";
-
-const getCompletedHiresData = async (userId: string) => {
-  // Get completed hires
-  const completedHires = await db.query.GigsTable.findMany({
-    where: and(
-      eq(GigsTable.buyerUserId, userId),
-      eq(GigsTable.statusInternal, "COMPLETED")
-    ),
-    with: { skillsRequired: { columns: { skillName: true } } },
-  });
-
-  // Count gigs per skill name and return array with name and value
-  const skillCounts: Record<string, number> = {};
-  completedHires.forEach((gig) => {
-    gig.skillsRequired.forEach((skill) => {
-      skillCounts[skill.skillName] = (skillCounts[skill.skillName] || 0) + 1;
-    });
-  });
-  const skillCountsArr = Object.entries(skillCounts).map(([name, value]) => ({
-    name,
-    value,
-  }));
-
-  const topSkills = skillCountsArr
-    .sort((a, b) => b.value - a.value)
-    .slice(0, 4);
-
-  return { completedHires, skillCountsArr, skillCounts, topSkills };
-};
-
-const getPaymentsData = async (userId: string) => {
-  // Calculate date 12 months ago
-  const twelveMonthsAgo = new Date();
-  twelveMonthsAgo.setMonth(twelveMonthsAgo.getMonth() - 12);
-
-  // Get last 12 months completed payments
-  const payments = await db.query.PaymentsTable.findMany({
-    where: and(
-      eq(PaymentsTable.payerUserId, userId),
-      eq(PaymentsTable.status, "COMPLETED"),
-      gt(PaymentsTable.createdAt, twelveMonthsAgo)
-    ),
-    columns: {
-      amountGross: true,
-      createdAt: true,
-    },
-  });
-
-  // Group payments into 4 groups of 3 months each
-  const now = new Date();
-  const groups: { amountGross: string; createdAt: Date }[][] = [[], [], [], []];
-
-  payments.forEach((payment) => {
-    const diffMonths =
-      (now.getFullYear() - payment.createdAt.getFullYear()) * 12 +
-      (now.getMonth() - payment.createdAt.getMonth());
-    const groupIndex = Math.floor(diffMonths / 3);
-    if (groupIndex >= 0 && groupIndex < 4) {
-      groups[groupIndex].push(payment);
-    }
-  });
-
-  // Calculate total expenses for each quarter, with clear naming
-  const barData = groups.map((group, idx) => {
-    const groupDate = new Date(now.getFullYear(), now.getMonth() - idx * 3, 1);
-    const quarter = Math.floor(groupDate.getMonth() / 3) + 1;
-    const year = groupDate.getFullYear().toString().slice(-2);
-    return {
-      name: `Q${quarter}'${year}`,
-      amount: group.reduce((sum, payment) => sum + Number(payment.amountGross), 0),
-    };
-  });
-
-  return barData.reverse();
-};
-
-const getBadgesData = async (userId: string) => {
-  // Get badges
-  const badges = await db
-    .select({
-      id: UserBadgesLinkTable.id,
-      awardedAt: UserBadgesLinkTable.awardedAt,
-      awardedBySystem: UserBadgesLinkTable.awardedBySystem,
-      notes: UserBadgesLinkTable.notes,
-      badge: {
-        id: BadgeDefinitionsTable.id,
-        name: BadgeDefinitionsTable.name,
-        description: BadgeDefinitionsTable.description,
-        icon: BadgeDefinitionsTable.iconUrlOrLucideName,
-        type: BadgeDefinitionsTable.type,
-      },
-    })
-    .from(UserBadgesLinkTable)
-    .innerJoin(
-      BadgeDefinitionsTable,
-      eq(UserBadgesLinkTable.badgeId, BadgeDefinitionsTable.id)
-    )
-    .where(eq(UserBadgesLinkTable.userId, userId));
-
-  const badgeDetails = badges.map((badge) => ({
-    id: badge.id,
-    name: badge.badge.name,
-    description: badge.badge.description,
-    icon: badge.badge.icon,
-    type: badge.badge.type,
-    awardedAt: badge.awardedAt,
-    awardedBySystem: badge.awardedBySystem,
-  }));
-
-  return badgeDetails;
-};
-
-const getReviewsData = async (userId: string) => {
-  const reviews = await db.query.ReviewsTable.findMany({
-    where: eq(ReviewsTable.targetUserId, userId),
-  });
-
-  const authorIds = reviews
-    .map((review) => review.authorUserId)
-    .filter((id): id is string => !!id);
-
-  const authors = await (async () => {
-    if (authorIds.length === 0) return [];
-    return await db.query.UsersTable.findMany({
-      where: (users, { inArray }) => inArray(users.id, authorIds),
-    });
-  })();
-
-  const authorsById = new Map(authors.map((author) => [author.id, author]));
-
-  const reviewsData = reviews.map((review) => {
-    const author = review.authorUserId
-      ? authorsById.get(review.authorUserId)
-      : null;
-    return {
-      id: review.id,
-      name: review.recommenderName || author?.fullName,
-      date: review.createdAt,
-      text: review.comment,
-    };
-  });
-
-  const totalReviews = reviews.length;
-
-  const positiveReviews = reviews.filter((item) => item.rating === 1).length;
-
-  const averageRating =
-    totalReviews > 0 ? (positiveReviews / totalReviews) * 100 : 0;
-
-  return { reviews: reviewsData, averageRating };
-};
+import { eq } from "drizzle-orm";
+import {
+  getCompletedHiresData,
+  getPaymentsData,
+  getBadgesData,
+  getReviewsData,
+} from "./buyer-profile-data";
 
 export const getGigBuyerProfileAction = async (
   token: string | undefined
@@ -220,113 +67,3 @@ export const getGigBuyerProfileAction = async (
   }
 };
 
-export const updateVideoUrlBuyerProfileAction = async (
-  videoUrl: string,
-  token?: string | undefined
-) => {
-  try {
-    if (!token) {
-      throw new Error("User ID is required to fetch buyer profile");
-    }
-
-    const { uid } = await isUserAuthenticated(token);
-    if (!uid) throw ERROR_CODES.UNAUTHORIZED;
-
-    const user = await db.query.UsersTable.findFirst({
-      where: eq(UsersTable.firebaseUid, uid),
-    });
-
-    if (!user) throw "User not found";
-
-    await db
-      .update(BuyerProfilesTable)
-      .set({
-        videoUrl: videoUrl,
-        updatedAt: new Date(),
-      })
-      .where(eq(BuyerProfilesTable.userId, user?.id));
-
-    return { success: true, data: "Url video updated successfully" };
-  } catch (error) {
-    console.log("Error saving video url", error);
-    return { success: false, data: "Url video updated successfully", error };
-  }
-};
-
-export const updateSocialLinkBuyerProfileAction = async (
-  socialLink: string,
-  token?: string
-): Promise<{ success: boolean; error?: string }> => {
-  try {
-    if (!token) {
-      return { success: false, error: "User token is required" };
-    }
-
-    const { uid } = await isUserAuthenticated(token);
-    if (!uid) return { success: false, error: "Unauthorized" };
-
-    const user = await db.query.UsersTable.findFirst({
-      where: eq(UsersTable.firebaseUid, uid),
-    });
-    if (!user) return { success: false, error: "User not found" };
-
-    await db
-      .update(BuyerProfilesTable)
-      .set({
-        socialLink,
-        updatedAt: new Date(),
-      })
-      .where(eq(BuyerProfilesTable.userId, user.id));
-
-    return { success: true };
-  } catch (error) {
-    console.error("Error saving social link", error);
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : "Unknown error",
-    };
-  }
-};
-
-type BusinessInfo = {
-  fullCompanyName: string;
-  location: { formatted_address: string; lat?: number; lng?: number };
-  companyRole: string;
-};
-
-export const updateBusinessInfoBuyerProfileAction = async (
-  { fullCompanyName, location, companyRole }: BusinessInfo,
-  token?: string
-): Promise<{ success: boolean; error?: string }> => {
-  try {
-    if (!token) {
-      return { success: false, error: "User token is required" };
-    }
-
-    const { uid } = await isUserAuthenticated(token);
-    if (!uid) return { success: false, error: "Unauthorized" };
-
-    const user = await db.query.UsersTable.findFirst({
-      where: eq(UsersTable.firebaseUid, uid),
-    });
-    if (!user) return { success: false, error: "User not found" };
-
-    await db
-      .update(BuyerProfilesTable)
-      .set({
-        fullCompanyName,
-        billingAddressJson: location,
-        companyRole,
-        updatedAt: new Date(),
-      })
-      .where(eq(BuyerProfilesTable.userId, user.id));
-
-    return { success: true };
-  } catch (error) {
-    console.error("Error saving business info", error);
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : "Unknown error",
-    };
-  }
-};
